IComp {TFARM}	R Documentation

Computes the Importance Index of a transcription factor in a set of association rules.

Description

Given an association rule and a transcription factor TFi, it is evaluated the contribution of TFi in the rule for the prediction of the presence of the item in the right-hand-side of the rule. Since this contribution is evaluated based on the variations of support and confidence of the rule, the user can visualize such variations by setting the parameter figures = TRUE.

Usage


IComp(TFi, rules_TF, rules_noTF, figures)

Arguments

TFi
string or string vector: the transcription factor (or combination of transcription factors) whose importance distribution is evaluated.

rules_TF
a set of rules in which TFi is present.

rules_noTF
a set of rules obtained from rules_TF removing the transcription factor (or combination of transcription factors) in TFi (obtained with the function rulesTF0).

figures
logical; if figures = TRUE, graphics with support and confidence distributions of the rules in rulesTF and rulesTF0 are returned.

Value

A list of four elements: the imp element is a vector of doubles with the importances of TFi in each rule in rulesTF; the delta element of the list is a list with variations of distributions of the two measures of support and confidence. This output is used in the function IPCA for the Principal Component Analysis of such distributions. The rwi element is a data.frame with the rules in rulesTF in which the transcription factor TFi is present and the rwo element is a data.frame with rules in rwi obtained removing the transcription factor TFi. Furthermore, if the input argument figures is set to TRUE, also the plots of the distributions of support and confidence of the rules before and after removing the transcription factor TFi are provided.

Examples


# Load r_FOSL2 and r_noFOSL2 from the data_man collection of datasets:
data('data_man')

# The Importance Indexes of FOSL2=1 in the set of rules r_FOSL2 are given by:
IComp('FOSL2=1', r_FOSL2, r_noFOSL2, figures=TRUE)

[Package TFARM version 1.1.1 Index]