IMP {TFARM}	R Documentation

Contains the mean Importance Index of each co-regulator.

Description

Within the data_man data collection, the dataset IMP has 4 columns and 12 rows: transcription factors present in at least one association rule are listed in the fist column (TF), the second column (imp) contains the means of the Importance Index associated with each trascription factor, the standard deviations for each transcription factor are reported in the third column (sd) and finally the number of rules found for each transcription factor are transcribed in the fourth column (nrules).

Usage


data("data_man")

Format

An object of class "data.frame"

Examples


# IMP is found in the data_man collection of datasets:
data("data_man")
head(IMP$imp)

[Package TFARM version 1.1.1 Index]