IMP_Z {TFARM}	R Documentation

Contains the Importance Index associated with each co-regulator which is present in at least one association rule.

Description

IMP_Z is a list of 12 elements, containig the Importance Indexes for all transcription factors (p) in the set of relevant association rules extracted. It is included in the data_man data collection.

Usage


data("data_man")

Format

An object of class "list"

Examples


# IMP_Z is found in the data_man collection of datasets:
data("data_man")
head(IMP_Z[[1]])

[Package TFARM version 1.1.1 Index]