I_c_2 {TFARM}	R Documentation

Contains the mean Importance Index of pairs of transcription factors which are present in at least one association rule.

Description

Within the data_man data collection, the dataset I_c_2 has 2 columns and 78 rows: the fist column (TF) contains the transcription factors (single or in pairs) and the second column (imp) lists the Importance Indexes associated with each trascription factor or pair of transcription factors.

Usage


data("data_man")

Format

An object of class "data.frame"

Examples


# I_c_2 is found in the data_man collection of datasets:
data("data_man")
head(I_c_2$imp)

[Package TFARM version 1.1.1 Index]