MCF7_chr1 {TFARM}	R Documentation

Contains genomic regions in the first chromosome of the MCF-7 human breast adenocarcinoma cell line at the ranges side, and the presence indexes of transcription factors in such regions at the metadata side.

Description

MCF7_chr1 is a Large GRanges in which metadata columns identify transcription factors and genomic coordinates of regions in the first chromosome of the MCF-7 human breast adenocarcinoma cell line are represented in the left-hand side of the GRanges, therefore each row is a different genomic region.

Usage


data("MCF7_chr1")

Format

An object of class "GRanges"

Examples


data("MCF7_chr1")
head(MCF7_chr1)

[Package TFARM version 1.1.1 Index]