TF_Imp {TFARM}	R Documentation

Contains the candidate co-regulators and the number of rules associated with them.

Description

Within the data_man data collection, the dataset TF_Imp has 3 columns and 12 rows: the fist column contains the transcription factors (IMP.TF), the Importance Indexes associated with each trascription factor are listed in the second column (IMP.imp) and the third column contains the number of rules found for each transcription factor (IMP.nrules).

Usage


data("data_man")

Format

An object of class "data.frame"

Examples


# TF_Imp is found in the data_man collection of datasets:
data("data_man")
head(TF_Imp$IMP.TF)

[Package TFARM version 1.1.1 Index]