heatI {TFARM}	R Documentation

Generates an heatmap visualization of the mean Importance Index of pairs of transcription factors.

Description

A square matrix (TFs x TFs) is built. The element (i,j) of such matrix contains the mean Importance Index (II) of the couple of transcription factors (TFi, TFj). This is rapresented in a heatmap visualization in which the scale color from blue to white indicates low mean importance of the couple, and from white to red high mean importance of the couple.

Usage


heatI(TFs, I)

Arguments

TFs
a string vector with names of transcription factors.

I
a vector of mean II of pairs of transcription factors in TFs.

Value

The Importance index of pairs of Transcription Factors as a heatmap

Examples


# Load p_TFs and I_c_2 from the data_man collection of datasets:
data('data_man')

# Heatmap visualization of the mean importances of transcription factors in p
# and their combinations in two elements:
heatI(p_TFs, I_c_2)

[Package TFARM version 1.1.1 Index]