imp_FOSL2 {TFARM}	R Documentation

Contains the set of Importance Indexes of FOSL2 in a given set of rules.

Description

Within the data_man data collection, imp_FOSL2 is a list of 4 elements, containing imp as a numeric vector in which the means of Importance Indexes are represented, delta as a data.frame in which the variations of support and confidence measures are reported, rwi as a data.frame of association rules where FOSL2 is present and rwo as a data.frame in which the same rules are considered for FOSL2=0.

Usage


data("data_man")

Format

An object of class "list"

Examples


# imp_FOSL2 is found in the data_man collection of datasets:
data("data_man")
head(imp_FOSL2$imp)

[Package TFARM version 1.1.1 Index]