items {TFARM}	R Documentation

Extracts items in the left-hand-side of an association rule.

Description

The function is used in the elaboration of the left-hand-side of association rules to search for the items.

Usage


items(itemset)

Arguments

itemset
object of class rules.

Value

A string vector with the items in the left-hand-side of the rule in itemset.

Examples


items('{TAF1=1,EP300=1,MAX=1}')
# the output is: 'TAF1=1','EP300=1','MAX=1'

[Package TFARM version 1.1.1 Index]