itemset {TFARM}	R Documentation

Builds the itemset in the left-hand-side of an association rule.

Description

The function is used in the construction of the left-hand-side of association rules to search for a rule of interest.

Usage


itemset(items)

Arguments

items
a string vector.

Value

Itemset of the left-hand-side of a searched rule, with the items in items.

Examples


itemset(c('TAF1=1','EP300=1'))
# the output is: 'TAF1=1,EP300=1'

[Package TFARM version 1.1.1 Index]