p_TFs {TFARM}	R Documentation

Contains co-regulators found in at least one association rule.

Description

p_TFs is a character vector that contains the transcription factors present in at least one association rule. It is included in the data_man data collection.

Usage


data("data_man")

Format

An object of class "character"

Examples


# p_TFs is found in the data_man collection of datasets:
data("data_man")
p_TFs

[Package TFARM version 1.1.1 Index]