presAbs {TFARM}	R Documentation

Splits a set of transcription factors in 'present' or 'absent' transcription factors in a set of rules.

Description

The function is used to find candidate transcription factor co-regulator with a given transcription factor, by finding the transcription factors that contribute to the prediction of the given trascription factor presence in a set of association rules.

Usage


presAbs(TFs, rules, type)

Arguments

TFs
a string vector: set of transcription factors to check their presence (TF=1) or absence (TF=0) in the set of rules.

rules
data.frame: rules and their quality measures.

type
logical parameter: if type = TRUE, only rules with all present transcription factors in the left-hand-side are considered (i.e., the left-hand-side of the extracted rules is for example TF1=1, TF2=1, TF3=1).

Value

A list of two string vectors: the list pres contains all the transcription factors in TFs that are present in rules, and the list abs contains all the transcription factors in TFs that are absent in rules.

Examples


library(GenomicRanges)
# Load r_TEAD4 from the data_man collection of datasets:
data('data_man')
# Load MCF7_chr1:
data('MCF7_chr1')

# Transcription factors present in at least one of the regions
# in the considered dataset:
c <- names(elementMetadata(MCF7_chr1))

names(presAbs(c, r_TEAD4, TRUE))

# Transcription factors present in at least one of the association rules:
p_TFs <- presAbs(c, r_TEAD4, TRUE)$pres
p_TFs

[Package TFARM version 1.1.1 Index]