r_TEAD4 {TFARM}	R Documentation

Contains the association rules for the prediction of the presence of the transcription factor TEAD4 in the considered genomic regions, i.e., with TEAD4 in the right-hand-side of the association rules.

Description

Within the data_man data collection, the dataset r_TEAD4 has 5 columns and 28 rows: the first column contains the left-hand-side of the rules (lhs), the second column is the right-hand-side of the rules (TEAD4=1) (rhs), the third column reports the support measures (support), the fourth column contains the confidence measures (confidence) and the lift measures are listed in the fifth column (lift).

Usage


data("data_man")

Format

An object of class "data.frame"

Examples


# Load r_TEAD4 from the data_man collection of datasets:
data("data_man")
head(r_TEAD4$lhs)

[Package TFARM version 1.1.1 Index]