rulesTF {TFARM}	R Documentation

Extracts a subset of rules that contain a certain transcription factor (or a combination of transcription factors) in their left-hand-side.

Description

From a set of relevant association rules, only the ones containing TFi in their left-hand-side are subsetted, together with their quality measures of support, confidence and lift. The function is then used for the evaluation of the importance distribution of TFi.

Usage


rulesTF(TFi, rules, verbose)

Arguments

TFi
a string with the name of the transcription factor (or combination of transcription factors) wanted in the left-hand-side of the rules to find.

rules
a data.frame with association rules and their quality measures of support, confidence and lift.

verbose
a logical parameter. If verbose = TRUE, a warning message is reported to the user when the set of rules containing TFi is empty.

Value

A data.frame with association rules containing TFi in their left-hand-side, with their quality measures of support, confidence and lift.

Examples


# Load r_TEAD4 from the data_man collection of datasets:
data('data_man')

r_FOSL2 <- rulesTF('FOSL2=1', r_TEAD4, verbose=FALSE)


[Package TFARM version 1.1.1 Index]