rulesTF0 {TFARM}	R Documentation

Substitutes the presence of a transcription factor (or a combination of transcription factors) in the left-hand-side of a set of rules, with its absence.

Description

The function substitutes the presence of a given transcription factor TFi (or a combination of transcription factors) chosen by the user with its absence, in the subset of relevant association rules extracted with the function rulesTF. Then it searches for the obtained rules and their quality measures of support, confidence and lift in the set of most relevant associations extracted with the function rulesGen. A rule is searched in all the association rules generable from the considered dataset using the function search_rule.

Usage


rulesTF0(TFi, sub_rules, all_rules, data, RHS)

Arguments

TFi
a string, or a string vector: transcription factor (or combination of transcription factors) to remove from the set of rules.

sub_rules
a data.frame with a subset of rules containing TFi, and their quality measures of support, confidence and lift (i.e., rules from which the user wants to remove TFi).

all_rules
a data.frame with a set of all the rules and their quality measures of support, confidence and lift, to be considered for the search of the obtained rules and their quality measures.

data
a GRanges object which contains the Indicator of presence matrix i.e., a matrix with 1 and 0 values representing presence or absence, respectively (in case other values different from 0 are present, all of them are considered as representing presence).

RHS
the right-hand-side of the considered association rules.

Value

A data.frame with all the rules in the set sub_rules in which the transcription factor (or combination of transcription factors) TFi is absent, and their quality measures of support, confidence and lift.

Examples


# Load r_TEAD4 and r_FOSL2 from the data_man collection of datasets:
data("data_man")
# Load MCF7_chr1:
data("MCF7_chr1")


r_noFOSL2 <- rulesTF0("FOSL2=1", r_FOSL2, r_TEAD4, MCF7_chr1, "TEAD4=1")


[Package TFARM version 1.1.1 Index]