search_rule {TFARM}	R Documentation

Finds an association rule in a dataset.

Description

The function looks for an association rule in a dataset data with a fixed left-hand-side (LHS) and a fixed right-hand-side (RHS). The function is used in the rulesTF0.

Usage


search_rule(data, LHS, RHS)

Arguments

data
a binary matrix or data.frame in which the rule is searched

LHS
a string with the left-hand-side of the searched rule

RHS
a string with the right-hand-side of the searched rule

Value

A vector with five elements: the left-hand-side of the searched rule, the right-hand-side of the searched rule, the support, confidence and lift of the found rule.

[Package TFARM version 1.1.1 Index]